libc: truncate memset fill value before wide pattern expansion

+++ libs/libc/string/lib_memset.c
@@ -58,7 +58,7 @@ FAR void *memset(FAR void *s, int c, size_t n)
    */
 
   uintptr_t addr  = (uintptr_t)s;
-  uint16_t  val16 = ((uint16_t)c << 8) | (uint16_t)c;
+  uint16_t  val16 = ((uint16_t)(uint8_t)c << 8) | (uint8_t)c;
   uint32_t  val32 = ((uint32_t)val16 << 16) | (uint32_t)val16;
 #ifdef CONFIG_MEMSET_64BIT
   uint64_t  val64 = ((uint64_t)val32 << 32) | (uint64_t)val32;
